allow buildpkey & buildwitness to be called from js code in addition to cli usage
